Move tests to a separated folder and add coverage
