+++ Makefile
@@ -10,7 +10,11 @@ stop: ## Stop the server
 	docker-compose down
 
 test: ## Test the application
-	docker-compose run app python3 -m unittest discover -s ./tests
+	docker-compose run app coverage run --omit=*/tests/* -m unittest discover
 
 test-verbose: ## Test the application in verbose mode
-	docker-compose run app python3 -m unittest discover -s ./tests -v
+	docker-compose run app coverage run -m unittest discover -v
+
+coverage-report: ## Run test with coverage
+	docker-compose run app coverage report
+
