+++ requirements.txt
@@ -1,2 +1,3 @@
 unittest-data-provider==1.0.1
 numpy==1.19.5
+coverage==5.4
