+++ src/.coveragerc
@@ -1,0 +1,5 @@
+# .coveragerc to control coverage.py
+
+[run]
+source=./
+omit=*/tests/*
