+++ src/tests/test_utils/print_board.py
@@ -1,7 +1,7 @@
 import sys
 from render import print_board
-from constant.ui import PRINT_BOARD_PLACE_MARBLE
+from constants.ui import PRINT_BOARD_PLACE_MARBLE
 
 def print_board_if_verbosity_is_set(board):
     if "-v" in sys.argv:
-        print_board(board, PRINT_BOARD_PLACE_MARBLE)
+        print_board(board, PRINT_BOARD_PLACE_MARBLE, [])
